requirements.txt
pillow>=9.0.0
opencv-python>=4.5.5
mysql-connector-python>=8.0.28
numpy>=1.22.0 